Add both about and shortcuts window in the help service
Both screens are alike. Now the about window is also rendered from a
Glade file.
