+++ gaphor/services/helpservice/about.glade
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated with glade 3.22.1 -->
+<interface>
+  <requires lib="gtk+" version="3.20"/>
+  <object class="GtkAboutDialog" id="about">
+    <property name="can_focus">False</property>
+    <property name="type_hint">dialog</property>
+    <property name="program_name">Gaphor</property>
+    <property name="version">1.1.0</property>
+    <property name="copyright" translatable="yes">Copyright © 2001-2019 Arjan J. Molenaar and Dan Yeaw</property>
+    <property name="comments" translatable="yes">Gaphor is the simple modeling tool written in Python</property>
+    <property name="website">https://github.com/gaphor/gaphor</property>
+    <property name="website_label" translatable="yes">Fork me on GitHub</property>
+    <property name="license" translatable="yes">This software is published under the terms of the
+Apache Software License, version 2.0.
+See the LICENSE.txt file for details.</property>
+    <property name="authors">Arjan Molenaar, Artur Wroblewski,
+Jeroen Vloothuis, Dan Yeaw,
+Enno Groeper, Adam Boduch,
+Alexis Howells, Melis Doğan</property>
+    <property name="translator_credits" translatable="yes">Jordi Mallach (ca), Antonin Delpeuch (fr), Ygor Mutti (pt_BR)</property>
+    <property name="logo">gaphor-96x96.png</property>
+    <child>
+      <placeholder/>
+    </child>
+    <child internal-child="vbox">
+      <object class="GtkBox">
+        <property name="can_focus">False</property>
+        <property name="orientation">vertical</property>
+        <property name="spacing">2</property>
+        <child internal-child="action_area">
+          <object class="GtkButtonBox">
+            <property name="can_focus">False</property>
+            <property name="layout_style">end</property>
+            <child>
+              <placeholder/>
+            </child>
+            <child>
+              <placeholder/>
+            </child>
+          </object>
+          <packing>
+            <property name="expand">False</property>
+            <property name="fill">False</property>
+            <property name="position">0</property>
+          </packing>
+        </child>
+        <child>
+          <placeholder/>
+        </child>
+      </object>
+    </child>
+  </object>
+</interface>
