+++ gaphor/services/helpservice/shortcuts.glade
